--- a/ghc-events.cabal
+++ b/ghc-events.cabal
@@ -49,7 +49,7 @@ library
   build-depends:    base       == 4.*,
                     containers >= 0.5 && < 0.6,
                     binary     >= 0.7 && < 0.9,
-                    bytestring >= 0.9.0,
+                    bytestring >= 0.10.4,
                     array      >= 0.2 && < 0.6
   exposed-modules:  GHC.RTS.Events,
                     GHC.RTS.EventsIncremental