@maoe 2017-02-12
Rebased onto the latest maoe/travis.