@kvelicka 2017-02-15
Looks promising! I'll have a proper look at it later this week.