@kvelicka 2017-02-16
Yup, looks really good indeed! I'm happy with this going into master, unless @Mikolaj / @simonmar object.