@maoe 2017-02-21
I fogot a whitespace in `inc` output. Pushed a fix for it. Let me know if it's better to squash it into 	23116c3.