@maoe 2017-02-24
Now that I got commit access, and @kvelicka and @Mikolaj seem to be happy with this patch. I'd like to merge it if @simonmar doesn't object.