@maoe 2017-03-01
I'm going to rebase this onto current master. There has been no objections so far. I think I'll merge this once CI passed after rebase.